requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.6
flask-sqlalchemy==3.1.1
flask-bcrypt==1.0.1
flask-jwt-extended==4.6.0
eventlet==0.35.1
torch
transformers
sentencepiece
protobuf
python-socketio[client]
